+++ components/inputs-mask/src/mask/index.tsx
@@ -7,6 +7,7 @@ import { useTestId } from "@gemunion/provider-test-id";
 
 export interface IMaskedInputProps {
   name: string;
+  required?: boolean;
   thousandSeparator?: string;
   valueIsNumericString?: boolean;
   prefix?: string;
@@ -39,6 +40,10 @@ export const MaskedInput: FC<IMaskedInputProps> = props => {
     form.setValue(props.name, formattedValue, { shouldTouch: true, shouldDirty: true });
   };
 
+  const onTriggerField = () => {
+    void form.trigger(name);
+  };
+
   return (
     <NumericFormat
       customInput={TextInput}
@@ -47,7 +52,8 @@ export const MaskedInput: FC<IMaskedInputProps> = props => {
       normalizeValue={normalizeValue}
       formatValue={formatValue}
       readOnly={readOnly}
-      onChange={() => {}}
+      onChange={onTriggerField}
+      onBlur={onTriggerField}
       {...rest}
       {...testIdProps}
     />