+++ providers/wallet/src/connectors/wallet-connect.ts
@@ -3,12 +3,31 @@ import { Web3ReactStore } from "@web3-react/types";
 import { WalletConnect } from "@web3-react/walletconnect-v2";
 
 // do not import from provider
-import { rpcUrls } from "../provider/constants";
+// import { rpcUrls } from "../provider/constants";
 
 /* javascript-obfuscator:disable */
 const WALLET_CONNECT_PROJECT_ID = process.env.WALLET_CONNECT_PROJECT_ID;
 // const WALLET_CONNECT_DEFAULT_CHAIN_ID = Number(process.env.WALLET_CONNECT_DEFAULT_CHAIN_ID);
 /* javascript-obfuscator:enable */
+
+// export const [walletConnect, hooks, store]: [WalletConnect, Web3ReactHooks, Web3ReactStore] =
+//   initializeConnector<WalletConnect>(
+//     (actions): WalletConnect =>
+//       new WalletConnect({
+//         actions,
+//         options: {
+//           projectId: WALLET_CONNECT_PROJECT_ID,
+//           chains: Object.keys(rpcUrls).map(Number), // [WALLET_CONNECT_DEFAULT_CHAIN_ID]
+//           optionalChains: Object.keys(rpcUrls).map(Number),
+//           showQrModal: true,
+//           qrModalOptions: {
+//             themeVariables: {
+//               "--wcm-z-index": "1500",
+//             },
+//           },
+//         },
+//       }),
+//   );
 
 export const [walletConnect, hooks, store]: [WalletConnect, Web3ReactHooks, Web3ReactStore] =
   initializeConnector<WalletConnect>(
@@ -17,8 +36,8 @@ export const [walletConnect, hooks, store]: [WalletConnect, Web3ReactHooks, Web3
         actions,
         options: {
           projectId: WALLET_CONNECT_PROJECT_ID,
-          chains: Object.keys(rpcUrls).map(Number), // [WALLET_CONNECT_DEFAULT_CHAIN_ID]
-          optionalChains: Object.keys(rpcUrls).map(Number),
+          chains: [1, 5, 97, 80002, 10001, 10000, 11155111, 13473], // [WALLET_CONNECT_DEFAULT_CHAIN_ID]
+          optionalChains: [1, 5, 97, 80002, 10001, 10000, 11155111, 13473],
           showQrModal: true,
           qrModalOptions: {
             themeVariables: {