+++ components/inputs-entity/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 7.3.84 (2024-07-20)
+
+### Bug Fixes
+
+- **deps:** update [@gemunion](https://github.com/gemunion) packages ([400f59b](https://github.com/gemunion/mui-packages/commit/400f59be08224b7c03146e3870002c141369cfbf))
+
 ## [7.3.83](https://github.com/gemunion/mui-packages/compare/@gemunion/mui-inputs-entity@7.3.82...@gemunion/mui-inputs-entity@7.3.83) (2024-07-16)
 
 ### Bug Fixes