+++ components/inputs-lab/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 5.0.115 (2024-07-20)
+
+**Note:** Version bump only for package @gemunion/mui-inputs-lab
+
 ## [5.0.114](https://github.com/gemunion/mui-packages/compare/@gemunion/mui-inputs-lab@5.0.113...@gemunion/mui-inputs-lab@5.0.114) (2024-07-16)
 
 **Note:** Version bump only for package @gemunion/mui-inputs-lab