+++ other/license-pages/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.3.7 (2024-07-20)
+
+### Bug Fixes
+
+- **deps:** update all non-major dependencies ([67302a5](https://github.com/gemunion/mui-packages/commit/67302a5fc5f9e83c3f784695eb8e5452edc3f0a6))
+
 ## [1.3.6](https://github.com/gemunion/mui-packages/compare/@gemunion/license-pages@1.3.5...@gemunion/license-pages@1.3.6) (2024-07-16)
 
 **Note:** Version bump only for package @gemunion/license-pages