+++ providers/picker/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 6.0.10 (2024-07-20)
+
+**Note:** Version bump only for package @gemunion/provider-picker
+
 ## [6.0.9](https://github.com/gemunion/mui-packages/compare/@gemunion/provider-picker@6.0.8...@gemunion/provider-picker@6.0.9) (2024-07-16)
 
 **Note:** Version bump only for package @gemunion/provider-picker